+++ src/main/java/io/stargate/sgv2/jsonapi/service/bridge/executor/ReactiveQueryExecutor.java
@@ -1,4 +1,4 @@
-package io.stargate.sgv3.docsapi.service.bridge.executor;
+package io.stargate.sgv2.jsonapi.service.bridge.executor;
 
 import com.google.protobuf.ByteString;
 import com.google.protobuf.BytesValue;
@@ -11,18 +11,18 @@ import io.stargate.sgv2.api.common.StargateRequestInfo;
 import io.stargate.sgv2.api.common.config.QueriesConfig;
 import io.stargate.sgv2.jsonapi.api.model.command.CommandResult;
 import io.stargate.sgv2.jsonapi.config.DocumentConfig;
-import io.stargate.sgv3.docsapi.service.sequencer.MultiQuerySequence;
-import io.stargate.sgv3.docsapi.service.sequencer.QueryOptions;
-import io.stargate.sgv3.docsapi.service.sequencer.QuerySequence;
-import io.stargate.sgv3.docsapi.service.sequencer.QuerySequenceSink;
-import io.stargate.sgv3.docsapi.service.sequencer.SingleQuerySequence;
-import io.stargate.sgv3.docsapi.service.sequencer.impl.CommandResultSink;
-import io.stargate.sgv3.docsapi.service.sequencer.impl.FunctionalQueryPipe;
-import io.stargate.sgv3.docsapi.service.sequencer.impl.FunctionalSink;
-import io.stargate.sgv3.docsapi.service.sequencer.impl.MultiQueryPipe;
-import io.stargate.sgv3.docsapi.service.sequencer.impl.MultiQuerySource;
-import io.stargate.sgv3.docsapi.service.sequencer.impl.SingleQueryPipe;
-import io.stargate.sgv3.docsapi.service.sequencer.impl.SingleQuerySource;
+import io.stargate.sgv2.jsonapi.service.sequencer.MultiQuerySequence;
+import io.stargate.sgv2.jsonapi.service.sequencer.QueryOptions;
+import io.stargate.sgv2.jsonapi.service.sequencer.QuerySequence;
+import io.stargate.sgv2.jsonapi.service.sequencer.QuerySequenceSink;
+import io.stargate.sgv2.jsonapi.service.sequencer.SingleQuerySequence;
+import io.stargate.sgv2.jsonapi.service.sequencer.impl.CommandResultSink;
+import io.stargate.sgv2.jsonapi.service.sequencer.impl.FunctionalQueryPipe;
+import io.stargate.sgv2.jsonapi.service.sequencer.impl.FunctionalSink;
+import io.stargate.sgv2.jsonapi.service.sequencer.impl.MultiQueryPipe;
+import io.stargate.sgv2.jsonapi.service.sequencer.impl.MultiQuerySource;
+import io.stargate.sgv2.jsonapi.service.sequencer.impl.SingleQueryPipe;
+import io.stargate.sgv2.jsonapi.service.sequencer.impl.SingleQuerySource;
 import java.util.ArrayList;
 import java.util.Base64;
 import java.util.List;
