+++ src/main/java/io/stargate/sgv2/jsonapi/service/operation/model/impl/DeleteOperation.java
@@ -8,10 +8,10 @@ import io.stargate.sgv2.jsonapi.service.bridge.serializer.CustomValueSerializers
 import io.stargate.sgv2.jsonapi.service.operation.model.ModifyOperation;
 import io.stargate.sgv2.jsonapi.service.operation.model.ReadOperation;
 import io.stargate.sgv2.jsonapi.service.operation.model.ReadOperation.FindResponse;
+import io.stargate.sgv2.jsonapi.service.sequencer.QueryOptions;
+import io.stargate.sgv2.jsonapi.service.sequencer.QuerySequence;
+import io.stargate.sgv2.jsonapi.service.sequencer.QuerySequenceSink;
 import io.stargate.sgv2.jsonapi.service.shredding.model.DocumentId;
-import io.stargate.sgv3.docsapi.service.sequencer.QueryOptions;
-import io.stargate.sgv3.docsapi.service.sequencer.QuerySequence;
-import io.stargate.sgv3.docsapi.service.sequencer.QuerySequenceSink;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
