+++ src/main/java/io/stargate/sgv2/jsonapi/service/operation/model/impl/FindOperation.java
@@ -12,9 +12,9 @@ import io.stargate.sgv2.jsonapi.exception.ErrorCode;
 import io.stargate.sgv2.jsonapi.exception.JsonApiException;
 import io.stargate.sgv2.jsonapi.service.bridge.serializer.CustomValueSerializers;
 import io.stargate.sgv2.jsonapi.service.operation.model.ReadOperation;
+import io.stargate.sgv2.jsonapi.service.sequencer.QuerySequence;
+import io.stargate.sgv2.jsonapi.service.sequencer.QuerySequenceSink;
 import io.stargate.sgv2.jsonapi.service.shredding.model.DocumentId;
-import io.stargate.sgv3.docsapi.service.sequencer.QuerySequence;
-import io.stargate.sgv3.docsapi.service.sequencer.QuerySequenceSink;
 import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
