+++ src/main/java/io/stargate/sgv2/jsonapi/service/operation/model/impl/InsertOperation.java
@@ -6,12 +6,12 @@ import io.stargate.sgv2.jsonapi.api.model.command.CommandContext;
 import io.stargate.sgv2.jsonapi.api.model.command.CommandResult;
 import io.stargate.sgv2.jsonapi.service.bridge.serializer.CustomValueSerializers;
 import io.stargate.sgv2.jsonapi.service.operation.model.ModifyOperation;
+import io.stargate.sgv2.jsonapi.service.sequencer.MultiQuerySequence;
+import io.stargate.sgv2.jsonapi.service.sequencer.QueryOptions;
+import io.stargate.sgv2.jsonapi.service.sequencer.QuerySequence;
+import io.stargate.sgv2.jsonapi.service.sequencer.QuerySequenceSink;
 import io.stargate.sgv2.jsonapi.service.shredding.model.DocumentId;
 import io.stargate.sgv2.jsonapi.service.shredding.model.WritableShreddedDocument;
-import io.stargate.sgv3.docsapi.service.sequencer.*;
-import io.stargate.sgv3.docsapi.service.sequencer.MultiQuerySequence;
-import io.stargate.sgv3.docsapi.service.sequencer.QueryOptions;
-import io.stargate.sgv3.docsapi.service.sequencer.QuerySequence;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.function.Supplier;
