+++ src/main/java/io/stargate/sgv2/jsonapi/service/sequencer/MultiQuerySequence.java
@@ -1,17 +1,18 @@
-package io.stargate.sgv3.docsapi.service.sequencer;
+package io.stargate.sgv2.jsonapi.service.sequencer;
 
 import io.stargate.bridge.proto.QueryOuterClass;
+import java.util.List;
 
 /**
- * Interface for a query sequence that executes a single query.
+ * Interface for a query sequence that executes multiple queries.
  *
  * @param <OUT> Type that is the outcome of this sequence part.
  */
-public interface SingleQuerySequence<OUT> extends QuerySequenceWithOptions<OUT> {
+public interface MultiQuerySequence<OUT> extends QuerySequenceWithOptions<List<OUT>> {
 
   /** Default handler, maps to the result itself, and (re-)throws throwable if any. */
   Handler<QueryOuterClass.ResultSet> DEFAULT_HANDLER =
-      (result, throwable) -> {
+      (result, throwable, i) -> {
         if (null != throwable) {
           throw throwable;
         }
@@ -20,14 +21,14 @@ public interface SingleQuerySequence<OUT> extends QuerySequenceWithOptions<OUT> 
 
   /** {@inheritDoc} */
   @Override
-  default SingleQuerySequence<OUT> withPageSize(int pageSize) {
+  default MultiQuerySequence<OUT> withPageSize(int pageSize) {
     QuerySequenceWithOptions.super.withPageSize(pageSize);
     return this;
   }
 
   /** {@inheritDoc} */
   @Override
-  default SingleQuerySequence<OUT> withPagingState(String pagingState) {
+  default MultiQuerySequence<OUT> withPagingState(String pagingState) {
     QuerySequenceWithOptions.super.withPagingState(pagingState);
     return this;
   }
@@ -41,7 +42,7 @@ public interface SingleQuerySequence<OUT> extends QuerySequenceWithOptions<OUT> 
    * @param handler New handler.
    * @return Updated instance.
    */
-  <T> SingleQuerySequence<T> withHandler(SingleQuerySequence.Handler<T> handler);
+  <T> MultiQuerySequence<T> withHandler(Handler<T> handler);
 
   /**
    * Query result handler.
@@ -52,12 +53,13 @@ public interface SingleQuerySequence<OUT> extends QuerySequenceWithOptions<OUT> 
   interface Handler<OUT> {
 
     /**
-     * Handles result of the query execution, or a throwable.
+     * Handles result of a single query execution, or a throwable.
      *
      * @param result Result of the query execution, can be <code>null</code>.
      * @param throwable Throwable being thrown during the execution of the query.
+     * @param index Index of the query to handle.
      * @return Result of the handler.
      */
-    OUT handle(QueryOuterClass.ResultSet result, Throwable throwable) throws Throwable;
+    OUT handle(QueryOuterClass.ResultSet result, Throwable throwable, int index) throws Throwable;
   }
 }
