+++ src/main/java/io/stargate/sgv2/jsonapi/service/sequencer/impl/EmptyQuerySequence.java
@@ -1,8 +1,8 @@
-package io.stargate.sgv3.docsapi.service.sequencer.impl;
+package io.stargate.sgv2.jsonapi.service.sequencer.impl;
 
 import io.smallrye.mutiny.Uni;
-import io.stargate.sgv3.docsapi.service.bridge.executor.ReactiveQueryExecutor;
-import io.stargate.sgv3.docsapi.service.sequencer.QuerySequence;
+import io.stargate.sgv2.jsonapi.service.bridge.executor.ReactiveQueryExecutor;
+import io.stargate.sgv2.jsonapi.service.sequencer.QuerySequence;
 
 /** Represents empty query sequence, started from no query. */
 public record EmptyQuerySequence() implements QuerySequence<Void> {
