+++ src/main/java/io/stargate/sgv2/jsonapi/service/sequencer/impl/SingleQuerySource.java
@@ -1,10 +1,10 @@
-package io.stargate.sgv3.docsapi.service.sequencer.impl;
+package io.stargate.sgv2.jsonapi.service.sequencer.impl;
 
 import io.smallrye.mutiny.Uni;
 import io.stargate.bridge.proto.QueryOuterClass;
-import io.stargate.sgv3.docsapi.service.bridge.executor.ReactiveQueryExecutor;
-import io.stargate.sgv3.docsapi.service.sequencer.QueryOptions;
-import io.stargate.sgv3.docsapi.service.sequencer.SingleQuerySequence;
+import io.stargate.sgv2.jsonapi.service.bridge.executor.ReactiveQueryExecutor;
+import io.stargate.sgv2.jsonapi.service.sequencer.QueryOptions;
+import io.stargate.sgv2.jsonapi.service.sequencer.SingleQuerySequence;
 
 /**
  * Represents start of the query sequence using a single query as source.
@@ -15,7 +15,7 @@ import io.stargate.sgv3.docsapi.service.sequencer.SingleQuerySequence;
  * @param <OUT> Output from this sequence part.
  */
 public record SingleQuerySource<OUT>(
-    QueryOuterClass.Query query, QueryOptions options, Handler<OUT> handler)
+    QueryOuterClass.Query query, QueryOptions options, SingleQuerySequence.Handler<OUT> handler)
     implements SingleQuerySequence<OUT> {
 
   /** {@inheritDoc} */
