+++ src/test/java/io/stargate/sgv2/jsonapi/service/operation/model/impl/FindOperationTest.java
@@ -13,9 +13,9 @@ import io.stargate.sgv2.jsonapi.api.model.command.CommandContext;
 import io.stargate.sgv2.jsonapi.api.model.command.CommandResult;
 import io.stargate.sgv2.jsonapi.service.bridge.AbstractValidatingStargateBridgeTest;
 import io.stargate.sgv2.jsonapi.service.bridge.ValidatingStargateBridge;
+import io.stargate.sgv2.jsonapi.service.bridge.executor.ReactiveQueryExecutor;
 import io.stargate.sgv2.jsonapi.service.bridge.serializer.CustomValueSerializers;
 import io.stargate.sgv2.jsonapi.service.shredding.model.DocumentId;
-import io.stargate.sgv3.docsapi.service.bridge.executor.ReactiveQueryExecutor;
 import java.util.List;
 import java.util.UUID;
 import java.util.function.Supplier;
